+++ vendor-service/src/main/java/com/fastline/vendorservice/application/VendorService.java
@@ -3,9 +3,10 @@ package com.fastline.vendorservice.application;
 import com.fastline.common.exception.CustomException;
 import com.fastline.common.exception.ErrorCode;
 import com.fastline.vendorservice.application.service.HubClient;
-import com.fastline.vendorservice.domain.entity.Order;
-import com.fastline.vendorservice.domain.entity.Product;
-import com.fastline.vendorservice.domain.entity.Vendor;
+import com.fastline.vendorservice.application.service.UserClient;
+import com.fastline.vendorservice.domain.model.Order;
+import com.fastline.vendorservice.domain.model.Product;
+import com.fastline.vendorservice.domain.model.Vendor;
 import com.fastline.vendorservice.domain.repository.VendorRepository;
 import com.fastline.vendorservice.domain.service.VendorOrderService;
 import com.fastline.vendorservice.domain.service.VendorProductService;
@@ -18,7 +19,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.stream.Collectors;
-
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -34,8 +34,9 @@ public class VendorService {
 	private final VendorOrderService vendorOrderService;
 
 	private final HubClient hubClient;
+	private final UserClient userClient;
 
-	public Vendor insert(VendorCreateRequest createRequest) {
+	public Vendor insert(VendorCreateRequest createRequest, Long userId) {
 
 		VendorType vendorType = VendorType.fromString(createRequest.type());
 
@@ -46,10 +47,15 @@ public class VendorService {
 						createRequest.roadName(),
 						createRequest.zipCode());
 
-        validateHubId(createRequest.hubId());
+		UUID passedHubId = validateHubId(createRequest.hubId());
 
-        Vendor vendor =
-				Vendor.create(createRequest.name(), vendorType, vendorAddress, createRequest.hubId());
+		Vendor vendor =
+				Vendor.create(createRequest.name(), vendorType, vendorAddress, passedHubId, userId);
+
+		UUID hubManagerId = userClient.getUserHubId(userId);
+		if (!isMasterUser(userId) && vendor.isHubManager(hubManagerId)) {
+			throw new CustomException(ErrorCode.VENDOR_FORBIDDEN);
+		}
 
 		return repository.insert(vendor);
 	}
@@ -62,40 +68,74 @@ public class VendorService {
 
 	@Transactional(readOnly = true)
 	public List<Product> findProductInVendor(UUID vendorId, Pageable pageable) {
+
 		return vendorProductService.findProductInVendor(vendorId, pageable);
 	}
 
 	@Transactional(readOnly = true)
-	public List<Order> findOrdersInVendor(UUID vendorId, Pageable pageable) {
+	public List<Order> findOrdersInVendor(UUID vendorId, Pageable pageable, Long userId) {
+
+		Vendor vendor = findByVendorId(vendorId);
+
+		UUID userHubId = userClient.getUserHubId(userId);
+		if (!isMasterUser(userId)
+				&& !vendor.isHubManager(userHubId)
+				&& !vendor.isVendorManager(userId)) {
+			throw new CustomException(ErrorCode.VENDOR_FORBIDDEN);
+		}
+
 		return vendorOrderService.findOrdersInVendor(vendorId, pageable);
 	}
 
-	public Vendor updateVendor(UUID vendorId, VendorUpdateRequest updateRequest) {
+	public Vendor updateVendor(UUID vendorId, VendorUpdateRequest updateRequest, Long userId) {
 
-		Vendor findVendor = repository.findByVendorId(vendorId);
-		if (updateRequest.hubId() != null && updateRequest.hubId() != findVendor.getHubId()) {
-            validateHubId(updateRequest.hubId());
+		Vendor vendor = repository.findByVendorId(vendorId);
+
+		UUID hubUserId = userClient.getUserHubId(userId);
+		if (!isMasterUser(userId)
+				&& !vendor.isHubManager(hubUserId)
+				&& !vendor.isVendorManager(userId)) {
+			throw new CustomException(ErrorCode.VENDOR_FORBIDDEN);
 		}
 
-		findVendor.update(updateRequest);
-		return repository.insert(findVendor);
+		if (updateRequest.hubId() != null && updateRequest.hubId() != vendor.getHubId()) {
+			validateHubId(updateRequest.hubId());
+		}
+
+		vendor.update(updateRequest);
+		return repository.insert(vendor);
 	}
 
-	public UUID deleteVendor(UUID vendorId) {
+	public UUID deleteVendor(UUID vendorId, Long userId) {
+
+		Vendor vendor = repository.findByVendorId(vendorId);
+
+		UUID userHubId = userClient.getUserHubId(userId);
+		if (!isMasterUser(userId) && !vendor.isHubManager(userHubId)) {
+			throw new CustomException(ErrorCode.VENDOR_FORBIDDEN);
+		}
+
 		return repository.deleteByVendorId(vendorId);
 	}
 
-    public List<UUID> findVendorHubId(UUID vendorSenderId, UUID vendorReceiverId) {
-        Map<UUID, UUID> mappingId = repository.findVendorHubId(vendorSenderId, vendorReceiverId).stream().collect(Collectors.toMap(VendorHubIdMappingObj::vendorID, VendorHubIdMappingObj::hubID));
+	public List<UUID> findHubIdInVendor(UUID vendorSenderId, UUID vendorReceiverId) {
+		Map<UUID, UUID> mappingId =
+				repository.findHubIdInVendor(vendorSenderId, vendorReceiverId).stream()
+						.collect(
+								Collectors.toMap(VendorHubIdMappingObj::vendorID, VendorHubIdMappingObj::hubID));
 
-        return List.of(mappingId.get(vendorSenderId), mappingId.get(vendorReceiverId));
-    }
+		return List.of(mappingId.get(vendorSenderId), mappingId.get(vendorReceiverId));
+	}
 
-    private boolean validateHubId(UUID hubId) {
+	private UUID validateHubId(UUID hubId) {
 
-        if (hubClient.validateHubId(hubId)) {
-            throw new CustomException(ErrorCode.VENDOR_HUBID_INVALIDATION);
-        }
-        return true;
-    }
+		if (hubClient.validateHubId(hubId)) {
+			throw new CustomException(ErrorCode.VENDOR_HUBID_INVALIDATION);
+		}
+		return hubId;
+	}
+
+	private boolean isMasterUser(Long userId) {
+		return userId == 1L;
+	}
 }
